Remove catch block in getting channelId

+++ handlers/dashboards/DashboardRefreshHandler.js
@@ -109,15 +109,17 @@ module.exports = {
 
 			dashboardSettings.forEach(async (dashboardSetting) => {
 				let guildId = dashboardSetting.guildId
-				let channel = await client.guilds.cache.get(guildId).channels.cache.get(dashboardSetting.dashboardChannelId).catch(() => {
-					//delete all dashboards belonging to a channel if the channel no longer exists
+				let channel = await client.guilds.cache.get(guildId).channels.cache.get(dashboardSetting.dashboardChannelId)
+					
+				//delete all dashboards belonging to a channel if the channel no longer exists
+				if(channel === undefined) {
 					Logger.info(`DashboardRefreshHandler.js`, 'refreshAll', 'The channel containing the dashboard no longer exists. Deleting all dashboards associated with the channel.')
-					DashboardSetting.destroy({
+					await DashboardSetting.destroy({
 						where: {
 							dashboardChannelId: dashboardSetting.dashboardChannelId
 						}
 					})
-				})
+				}
 				
 				let message = await channel.messages.fetch(dashboardSetting.dashboardMessageId).catch(() => {
 					//delete dashboard if its bound message no longer exists
